requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
passlib[bcrypt]
python-dotenv
httpx
